Renamed test dom3->dom2
