+++ tests/dom/Makefile
@@ -1,3 +1,3 @@
 include ../tests.make
 
-all: dom1.js dom3.js
+all: dom1.js dom2.js
