--- tests/dom/test3.html
@@ -1,7 +1,0 @@
-<html>
-<head>
-	<script src="dom3.js"></script>
-</head>
-<body>
-</body>
-</html>
